Allow Ctrl+S on file renaming on 'keyboard' mode + hide 'alphabet switch' prompt when required
Ctrl+S used to work only on controller input, because the code handling it is only active when the keyboard is focused. The alphabet switch thing is a nitpick about the 'press Tab to witch between hiragana and katakana' prompt, that shouldn't display on keyboard mode, because it only works (and is only relevant) in controller mode.