Fix more sources of memory leaks
Namely, alarms that can store references to levels after using a Badeline boost, and some mod entities keeping themselves or some of their components in static variables. For the second one, we are dissociating the level and all its entities, so that even if the entity is held by some static variable, the level can be cleaned up.