Hook NLua to redirect calls to Assembly.GetTypes() to Extensions.GetTypesSafe()
As discussed on Discord long ago, this fixes crashes with NLua (so Lua Cutscenes) when mods with 'optional dependencies' are installed.
We can't patch NLua directly because it would create a circular dependency on Everest => NLua => Everest.
This fixes, for example, skipping the intro cutscene of the Spring Collab when Guneline is installed but not CelesteNet (the cutscene is broken on Windows, and crashes the game on Linux).