Fix 'multiple options selected on main menu' issue (again)
Calling RebuildMainAndTitle() multiple times in a row causes the issue. And this happens by just opening Mod Options because the settings are reloaded (so all settings are set, and some of them trigger a menu rebuild). Reloading mod settings upon entering Mod Options doesn't seem to be useful unless I'm missing something, so I commented that out. (OnInputInitialize was only necessary because of the reload, so I removed it as well.)