Add checks to be sure we don't restore backup mod save data when unnecessary
Checking if no mod level set is present in the file is not enough, as it causes issues when the player has no custom maps installed.
Checking for a flag causes problems to people that downgrade then upgrade, with the save getting sent 'back in time'.
Checking for both the flag and mod level sets though...