Show messages when mods start loading in dependency downloader
When you're missing Clutter Helper for the Spring Collab and you hit 'install dependencies', you saw 'installing ClutterHelper' for quite a while, because after Clutter Helper is installed, the Spring Collab is loaded. The console didn't show that the Spring Collab was loading at all. And if something goes wrong when loading the collab, it would look like it is Clutter Helper's fault...