Warn if mod assembly doesn't contain an EverestModule
Log a warning if the assembly of a code mod doesn't contain an EverestModule. All code mods should have one, and the ones who don't can bypass this warning by creating an EverestModule which either does nothing or subclasses NullModule. This can help developers catch errors when they either break the module, or it somehow fails to load (in which case it's not returned by asm.GetTypesSafe()).