Backwards-compat for mods that overwrite the sound index for 'o'
Some existing mods (e.g. Into the Jungle) previously got away with overwriting the sound index for 'o' (it's unused in the vanilla ForegroundTiles.xml but is assigned in the LevelLoader patch). Not allowing them to do this creates more issues than it solves, so just allow it via special case