Improve assembly resolution
With the original implementation, every time an assembly couldn't be resolved, we searched through every file of every loaded mod to find a matching assembly. This adds some major extra load times when the tracker is initializing + any other time we need to get the list of types since every time we can't resolve something we're unzipping every archive and searching every folder.

Instead of that, we can use the mod content maps, since they already have a list of every file in the mod folders. This maintains the current load order but removes the duplicate crawling.