Make logging less "verbose"
1. Changes default threshold for logging to Info.
2. Changes detour log level to Debug.

This makes logs much easier to read. Log levels for individual tags can still be set in mod settings, but the ones for Everest that are useful for debugging crashes should all be Info or higher already. Detour logs can be printed on demand with the "logdetours" command, so we're not really losing functionality there.

I'll probably also work on a logging page in the wiki so we can define some good practices.