Change default minimum log level to a setting
This lets you lower the threshold for Everest + all tags with unspecified log levels, rather than forcing you to add each individual one to the LogLevels dictionary.

Also edited LogDetours to use the level of the setting if the "logdetours" command is used, since in that case we want it to always print.