Apply scaling to other properties
Applies "scale" as the first property, and then adjusts offsets, radii, etc. by the scale for compatibility. Without this, the scale attribute breaks pretty much all of these properties.

These problems existed before with Ahorn/Lonn scaling, but this feels a bit more urgent now that there's an Everest property for it.