Handle duplicate room names
The old leveldata retrieval method just had a list of all rooms and gave you the first one that matched. The new cache system has a dictionary of room names, which broke with duplicate names. This change tweaks the behavior to only add the first instance of a room name and log a warning for any duplicates.

Also removed a check that regenerated the cache if the room count changed, which triggered a regen for every access when a duplicate room name was involved (also I don't think it's needed in general)