Make GetTypesSafe filter out "unsafe" types
Currently, a lot of issues are emerging where code relying on `GetTypesSafe` fails in some way or another, because it directly or indirectly depends on a type contained in an unloaded optional dependency, as these will be contained in the output of `GetTypesSafe` and subsequently crash when specific reflection operations are performed. This PR would fix these kinds of issues at the source, as it would ensure all types returned by `GetTypesSafe` have been "tested" for dependencies on non-present assemblies.

If this were to be merged, https://github.com/EverestAPI/Everest/commit/f479dbbbe9451635d1b1fd3f55ad0322baa39e11 can be reverted.