Add RestartedFromGolden field to Session
Field will be true for a session created from a golden berry restart.

Dashless goldens and similar challenges usually rely on Session.StartedFromBeginning, which is false after a golden restart if the golden is not in the first room. This new field lets you account for that instead of requiring the player to restart the chapter.

The dashless golden handler was patched to use this new field.

The IL patches that were edited have also been reworked to use ILCursor. One side effect is that many invalid strings that were saved in _LoadStrings have been removed. Apart from that, the generated IL is identical.