Ensure dream blocks have the same number of particles on all platforms
Windows can sometimes give a different value to macOS for float calculations close to a whole number, whereas they both agree if calculating with doubles.
Due to this, DreamBlock.Setup calculates a different number of particles for certain dream blocks.  This means more random samples taken, and a desync later in the room.
Solution is to subtract a small value before truncating to int, which gives the desired value on both platforms.