Make the debug camera controls also control the slight camera movement on the map.
The slight camera rotation on the map is hardcoded to use WASD. This can get annoying for people who have menu buttons bound to WASD. This changes those buttons to be set by the debug camera controls in the Everest mod settings. It also adds default controller controls for the camera movement so that it will still behave like normal if you don't change settings from default.