Clarify build steps in README
When the building instructions were previously modified to include
cloning submodules, the words "or by downloading the .zip from
Github" were left in place. Building Everest in this way is currently
not supported because these source distributions do not include the
required submodules.

Also clarified required versions of the .NET SDK.

Fixes #718.